move various helpers
